requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
libsql-experimental==0.0.55
playwright==1.44.0
apscheduler==3.10.4
jinja2==3.1.4
python-multipart==0.0.9
itsdangerous==2.2.0
pydantic-settings==2.2.1
httpx==0.27.0
Pillow==10.3.0
boto3==1.34.0
aiofiles==23.2.1
